move decoder to utils
